third_party/WebKit/LayoutTests/fast/forms/reset-autofilled-expected.txt

This tests that a field's autofilled state is reverted when the form is reset. It can only be run using the test harness.
    
PASS countAutofilledElements() is 3
PASS countAutofilledElements() is 0
